chore: Modularization strategy
The main reason to create all the different Gradle modules
is for scalability, separation of concerns, better testing
among other benefits.

Here I define the relation between the modules.

:app the entry point of the app depends on :feature-users
:feature-users the UI depends on :data-repository
:data-repository depends on :data-remote

And the :core module that all depends on it, acts as the
abstraction module. That is, the UI doesn't know anything
about how the data is retrieved.

The dependency inversion principles is follow here by depending
on interfaces instead of concrete functions.