Get user list on repository layer
We have two different data classes to handle the data divided in two
blocks:
1.
 - UserApiModel. Contains the data from the server.
 - User. Contains the data that was parsed to be ready to display
  on the UI.

2.
 - DataState. Contains the data from the server mapped as UI model
and wrapped inside the DataState sealed class.
 - UiState. Contains the relevant data for the UI.

The block 2 is a wrapped that contains the data of the block 1 plus
other state of the network and UI.

In this each module doesn't depend on each other data. So it is
easier for future changes and also avoid possible errors or dummy
data fields.

The result from the server is mapped and the flow emit new values
to the UI layer.