feat: Show user detail view
IMPORTANT!!!

There is no endpoint on the API to retrieve a single user.
So, the data on the user list should also contain all the detail
data (not yet implemented).
This is a really bad behaviour as a lot of data needs to be
retrieved when it isn't necessary, wasting resources.

Also, the data displayed on the detail view needs to be send from
the user list. I opt for send all this data using the navigation arguments.

But this is consider an anti-pattern.

The proper way would be collect the required data on each view.
And pass to the detail view only the user id.
That is, the single user view ideally would make a request to the
server to get the data of the user clicked, but this endpoint is
not available for now.